backend/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-socketio>=5.3
python-dotenv>=1.0
eventlet>=0.35
yfinance>=0.2
requests>=2.31
langchain>=0.2
langchain-openai>=0.1
langchain-community>=0.2
serpapi>=0.1
wikipedia>=1.4
beautifulsoup4>=4.12
psutil>=5.9
